% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global-solution-methods.R
\name{den_haan_marcet}
\alias{den_haan_marcet}
\title{Den Haan-Marcet accuracy statistic}
\usage{
den_haan_marcet(x, ...)
}
\arguments{
\item{x}{A solved policy: a \code{GlobalSolution} or a perturbation
\code{DecisionRules*} object.}

\item{...}{Method arguments; see
\code{\link{den_haan_marcet.GlobalSolution}}.}
}
\description{
Den Haan-Marcet accuracy statistic
}
