% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hank-egm2.R
\name{.hank2_apply_coord_rows}
\alias{.hank2_apply_coord_rows}
\title{Apply interpolation coordinates to PER-ROW knot values}
\usage{
.hank2_apply_coord_rows(i, p, Y)
}
\description{
Row-wise \code{yq[l, ] = p[l, ] * Y[l, i[l, ]] + (1-p[l, ]) * Y[l, i[l, ]+1]}.
}
\keyword{internal}
