% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hank-distribution2.R
\name{.hank2_vec_to_arr}
\alias{.hank2_vec_to_arr}
\title{Inverse of \code{\link{.hank2_arr_to_vec}}}
\usage{
.hank2_vec_to_arr(v, n_e, n_b, n_a)
}
\arguments{
\item{v}{Numeric length-\code{n_e*n_b*n_a} vector in distribution order.}

\item{n_e, n_b, n_a}{Integer dimensions.}
}
\value{
Numeric \code{n_e x n_b x n_a} array.
}
\description{
Inverse of \code{\link{.hank2_arr_to_vec}}
}
\keyword{internal}
