% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hank-aggregate-risk.R
\name{.hank_agg_policy_at}
\alias{.hank_agg_policy_at}
\title{Interpolate a stacked policy array in log K at one aggregate state}
\usage{
.hank_agg_policy_at(P, log_K_grid, log_K, z, n_e, n_a, n_k)
}
\arguments{
\item{P}{Numeric \code{(n_e*n_a) x (n_k*n_z)} policy array.}

\item{log_K_grid}{Increasing \code{log(K_grid)}.}

\item{log_K}{Scalar query in logs.}

\item{z}{Integer aggregate-state index.}

\item{n_e, n_a}{Household grid dimensions.}

\item{n_k}{Number of aggregate-capital nodes.}
}
\value{
An \code{n_e x n_a} matrix.
}
\description{
Interpolate a stacked policy array in log K at one aggregate state
}
\keyword{internal}
