% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hank-model.R
\name{.hank_block_jacobian}
\alias{.hank_block_jacobian}
\title{Block Jacobian dispatch (simple: analytic or FD; het: fake-news)}
\usage{
.hank_block_jacobian(blk, ss, T_h)
}
\description{
Block Jacobian dispatch (simple: analytic or FD; het: fake-news)
}
\keyword{internal}
