% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hank-income.R
\name{.hank_check_eun_rates}
\alias{.hank_check_eun_rates}
\title{Validate the six E/U/N employment-margin transition rates}
\usage{
.hank_check_eun_rates(f_ue, s_eu, p_un, p_nu, f_ne, s_en, prefix = "")
}
\description{
Validate the six E/U/N employment-margin transition rates
}
\keyword{internal}
