% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hank-jacobian2d.R
\name{.hank_curly_sweep2d}
\alias{.hank_curly_sweep2d}
\title{Backward sweep for the discrete-adjustment fake-news algorithm}
\usage{
.hank_curly_sweep2d(block, T_h, i, outputs, delta_in, delta_va, delta_d)
}
\description{
Backward sweep for the discrete-adjustment fake-news algorithm
}
\keyword{internal}
