% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hank-jacobian2.R
\name{.hank_het2_check_outputs}
\alias{.hank_het2_check_outputs}
\title{Validate requested two-asset het-block outputs}
\usage{
.hank_het2_check_outputs(outputs)
}
\description{
Validate requested two-asset het-block outputs
}
\keyword{internal}
