% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hank-jacobian2d.R
\name{.hank_het2d_check_inputs}
\alias{.hank_het2d_check_inputs}
\title{Validate requested discrete-adjustment het-block Jacobian inputs}
\usage{
.hank_het2d_check_inputs(block, inputs)
}
\description{
Validate requested discrete-adjustment het-block Jacobian inputs
}
\keyword{internal}
