% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hank-jacobian2d.R
\name{.hank_het2d_check_outputs}
\alias{.hank_het2d_check_outputs}
\title{Validate requested discrete-adjustment het-block outputs}
\usage{
.hank_het2d_check_outputs(outputs)
}
\description{
Validate requested discrete-adjustment het-block outputs
}
\keyword{internal}
