% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hank-jacobian.R
\name{.hank_het_is_flow_output}
\alias{.hank_het_is_flow_output}
\title{Is a het-block output name a flow ("F_xy") output?}
\usage{
.hank_het_is_flow_output(o)
}
\description{
Is a het-block output name a flow ("F_xy") output?
}
\keyword{internal}
