% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hank-jacobian.R
\name{.hank_het_state_indicator}
\alias{.hank_het_state_indicator}
\title{Fixed indicator vector for a het-block employment-margin STATE output}
\usage{
.hank_het_state_indicator(idx, n_e, n_a)
}
\description{
Fixed indicator vector for a het-block employment-margin STATE output
}
\keyword{internal}
