% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hank-reweighting.R
\name{.hank_mixture_check_omega}
\alias{.hank_mixture_check_omega}
\title{Validate a mixture-weight vector against a list of blocks}
\usage{
.hank_mixture_check_omega(blocks, omega)
}
\description{
Validate a mixture-weight vector against a list of blocks
}
\keyword{internal}
