% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hank-reweighting.R
\name{.hank_reweight_functional_cov}
\alias{.hank_reweight_functional_cov}
\title{Proper K x K covariance of the projected net reweighting}
\usage{
.hank_reweight_functional_cov(B, D0, D1, N)
}
\description{
Proper K x K covariance of the projected net reweighting
}
\keyword{internal}
