% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hank-model.R
\name{.hank_simple_jac_fd}
\alias{.hank_simple_jac_fd}
\title{Finite-difference sequence-space Jacobian of a simple block}
\usage{
.hank_simple_jac_fd(blk, ss, T_h, delta = 1e-06)
}
\description{
Finite-difference sequence-space Jacobian of a simple block
}
\keyword{internal}
