% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hank-estimation.R
\name{.hank_stacked_cov}
\alias{.hank_stacked_cov}
\title{Block-Toeplitz covariance of a stacked aggregate sample}
\usage{
.hank_stacked_cov(G, T_data, me_variance = 0)
}
\arguments{
\item{G}{Autocovariance array from \code{\link{hank_autocov}} (max lag must
be >= \code{T_data - 1}).}

\item{T_data}{Integer: number of time periods in the sample.}

\item{me_variance}{Numeric >= 0: measurement-error variance added on the diagonal
(required when n_obs > n_shocks to avoid singularity).}
}
\value{
A \code{(T_data*n_obs) x (T_data*n_obs)} covariance matrix, stacked
  time-major (row \code{(t-1)*n_obs + j}).
}
\description{
Block-Toeplitz covariance of a stacked aggregate sample
}
\keyword{internal}
