% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hank-diagnostics.R
\name{.hank_symbol_eval}
\alias{.hank_symbol_eval}
\title{Evaluate a matrix symbol on the unit circle by FFT}
\usage{
.hank_symbol_eval(jk, tau, N)
}
\arguments{
\item{jk}{Array \code{n x n x (2 tau + 1)} from
\code{\link{.hank_symbol_blocks}}.}

\item{tau}{Band half-width.}

\item{N}{Number of nodes \eqn{z_m = e^{2\pi i m/N}}, \eqn{m = 0..N-1}
(counter-clockwise); needs \code{N > 2 tau}.}
}
\value{
Complex array \code{N x n x n} with \code{[m + 1, , ] = j(z_m)}.
}
\description{
Evaluate a matrix symbol on the unit circle by FFT
}
\keyword{internal}
