% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hank-model.R
\name{.hank_topo_order}
\alias{.hank_topo_order}
\title{Topologically order blocks by input/output dependencies}
\usage{
.hank_topo_order(blocks, sources)
}
\description{
Topologically order blocks by input/output dependencies
}
\keyword{internal}
