% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hank-distribution.R
\name{.hank_vec_to_mat}
\alias{.hank_vec_to_mat}
\title{Reshape a distribution-order vector back to an \code{n_e x n_a} matrix}
\usage{
.hank_vec_to_mat(v, n_e, n_a)
}
\arguments{
\item{v}{A length-\code{n_e*n_a} vector in distribution order.}

\item{n_e, n_a}{Grid dimensions.}
}
\value{
An \code{n_e x n_a} matrix.
}
\description{
Reshape a distribution-order vector back to an \code{n_e x n_a} matrix
}
\keyword{internal}
