% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obc-regime-path.R
\name{.obc_binding_to_bitfield}
\alias{.obc_binding_to_bitfield}
\title{Encode a logical binding matrix as an integer regime-bitfield}
\usage{
.obc_binding_to_bitfield(binding)
}
\arguments{
\item{binding}{A \code{T x n_constr} logical matrix.}
}
\value{
Integer vector of length \code{T}.
}
\description{
Encode a logical binding matrix as an integer regime-bitfield
}
\keyword{internal}
