% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obc-regime-path.R
\name{.obc_bitfield_to_binding}
\alias{.obc_bitfield_to_binding}
\title{Decode an integer regime-bitfield to a logical binding matrix}
\usage{
.obc_bitfield_to_binding(codes, n_constr)
}
\arguments{
\item{codes}{Integer vector (length T): bit \code{b-1} set means constraint
\code{b} binds in that period.}

\item{n_constr}{Number of constraints.}
}
\value{
A \code{T x n_constr} logical matrix.
}
\description{
Decode an integer regime-bitfield to a logical binding matrix
}
\keyword{internal}
