% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag-plot-theme.R
\name{dynhr_na_colour}
\alias{dynhr_na_colour}
\title{Colour used for missing / invalid / out-of-sample data}
\usage{
dynhr_na_colour
}
\description{
Colour used for missing / invalid / out-of-sample data
}
