% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag-plot-theme.R
\name{dynhr_palette_light}
\alias{dynhr_palette_light}
\title{Default dynhr fill palette (Tol light, aligned with vibrant order)}
\usage{
dynhr_palette_light
}
\description{
Default dynhr fill palette (Tol light, aligned with vibrant order)
}
