% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag-plot-theme.R
\name{dynhr_palette_vibrant}
\alias{dynhr_palette_vibrant}
\title{Default dynhr line colour palette (Tol vibrant in dynhr order)}
\usage{
dynhr_palette_vibrant
}
\description{
Default dynhr line colour palette (Tol vibrant in dynhr order)
}
