% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag-plot-theme.R
\name{dynhr_primary_colour}
\alias{dynhr_primary_colour}
\title{Primary brand colour for headings and accents (vibrant blue)}
\usage{
dynhr_primary_colour
}
\description{
Primary brand colour for headings and accents (vibrant blue)
}
