% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{dynhr_reset_options}
\alias{dynhr_reset_options}
\title{Reset dynhr options to package defaults}
\usage{
dynhr_reset_options(...)
}
\arguments{
\item{...}{Character names of options to reset. If empty, resets all.}
}
\description{
Reset dynhr options to package defaults
}
