% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{estimate_mode}
\alias{estimate_mode}
\title{Run mode-finding for a solved DSGE model}
\usage{
estimate_mode(...)
}
\arguments{
\item{...}{Arguments forwarded to \code{\link{run_mode_finding}}.}
}
\value{
A \code{dynhr_mode_result} object.
}
\description{
Convenience alias for \code{\link{run_mode_finding}}.
}
\seealso{
\code{\link{run_mode_finding}}, \code{\link{run_posterior_estimation}}
}
