% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global-solution-methods.R
\name{euler_errors}
\alias{euler_errors}
\title{Euler equation errors generic}
\usage{
euler_errors(x, ...)
}
\arguments{
\item{x}{An object with a global solution (e.g., \code{GlobalSolution}).}

\item{...}{Additional arguments.}
}
\description{
Euler equation errors generic
}
