% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hank-estimation.R
\name{hank_autocov}
\alias{hank_autocov}
\title{Autocovariance function from MA coefficients}
\usage{
hank_autocov(Theta, sigma_eps, n_lags)
}
\arguments{
\item{Theta}{\code{T_h x n_obs} MA coefficients (see
\code{\link{hank_ma_coefficients}}).}

\item{sigma_eps}{Numeric: standard deviation of the (scalar) shock innovation.}

\item{n_lags}{Integer: maximum lag \eqn{k} to return.}
}
\value{
An \code{n_obs x n_obs x (n_lags+1)} array with slice \code{k+1}
  giving \eqn{\Gamma(k)}.
}
\description{
Autocovariance function from MA coefficients
}
