% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hank-jacobian2d.R
\name{hank_het2d_jacobian_nd}
\alias{hank_het2d_jacobian_nd}
\title{Brute-force ND Jacobian of a discrete-adjustment het block}
\usage{
hank_het2d_jacobian_nd(
  block,
  T_h,
  inputs = c("rb", "ra", "w"),
  outputs = c("B", "A", "C"),
  delta = 1e-05
)
}
\arguments{
\item{block}{A \code{\link{hank_het2d_block}}.}

\item{T_h}{Horizon.}

\item{inputs}{Subset of \code{c("rb", "ra", "w", "Tr")}.}

\item{outputs}{Subset of \code{c("B", "A", "C", "CHI", "ADJ")}.}

\item{delta}{FD step.}
}
\value{
Nested list \code{J[[output]][[input]]} of \code{T_h x T_h}
  matrices.
}
\description{
Reference Jacobian by central differences of
\code{\link{hank_td2d_nonlinear}} -- \eqn{O(T)} full transitions per
(input, date), the mandatory oracle for \code{\link{hank_het2d_jacobian}}.
}
\seealso{
\code{\link{hank_het2d_jacobian}}
}
\keyword{internal}
