% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hank-diagnostics.R
\name{hank_plot_jacobian}
\alias{hank_plot_jacobian}
\title{Heatmap of a sequence-space Jacobian block}
\usage{
hank_plot_jacobian(Jblock, main = "Sequence-space Jacobian", ...)
}
\arguments{
\item{Jblock}{A \code{T x T} Jacobian matrix (e.g. \code{J[["C"]][["r"]]}).}

\item{main}{Plot title.}

\item{...}{Passed to \code{graphics::image}.}
}
\value{
Invisibly, \code{Jblock}.
}
\description{
Heatmap of a sequence-space Jacobian block
}
