% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hank-truncation.R
\name{hank_reiter_irf}
\alias{hank_reiter_irf}
\title{Impulse responses of a Reiter state space}
\usage{
hank_reiter_irf(rss, T_h = 100L, shock = 0.01)
}
\arguments{
\item{rss}{A \code{\link{hank_reiter_statespace}} object.}

\item{T_h}{Horizon.}

\item{shock}{Innovation size (log-TFP units) hitting \code{dz} at t = 1.}
}
\value{
A \code{T_h x n_obs} matrix of observable paths (columns named as
  \code{rss$obs_names}).
}
\description{
Impulse responses of a Reiter state space
}
