% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochsimul-monolith.R
\name{print.DecisionRules}
\alias{print.DecisionRules}
\title{Print first-order decision rules}
\usage{
\method{print}{DecisionRules}(x, ...)
}
\arguments{
\item{x}{A \code{DecisionRules} object from [solve_perturbation()].}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print first-order decision rules
}
