% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve-perturbation-order2.R
\name{print.DecisionRules2}
\alias{print.DecisionRules2}
\title{Print method for second-order decision rules}
\usage{
\method{print}{DecisionRules2}(x, ...)
}
\arguments{
\item{x}{A \code{DecisionRules2} object.}

\item{...}{Unused; included for S3 compatibility.}
}
\description{
Print method for second-order decision rules
}
