% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve-perturbation-order4.R
\name{print.DecisionRules4}
\alias{print.DecisionRules4}
\title{Print method for 4th-order decision rules}
\usage{
\method{print}{DecisionRules4}(x, ...)
}
\arguments{
\item{x}{A \code{DecisionRules4} object.}

\item{...}{Unused; included for S3 compatibility.}
}
\description{
Print method for 4th-order decision rules
}
