% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve-perturbation-order5.R
\name{print.DecisionRules5}
\alias{print.DecisionRules5}
\title{Print method for 5th-order decision rules}
\usage{
\method{print}{DecisionRules5}(x, ...)
}
\arguments{
\item{x}{A \code{DecisionRules5} object.}

\item{...}{Unused; included for S3 compatibility.}
}
\description{
Print method for 5th-order decision rules
}
