% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag-deep-helpers.R
\name{print.dynhr_deep_spec}
\alias{print.dynhr_deep_spec}
\title{Print a deep-parameter taxonomy}
\usage{
\method{print}{dynhr_deep_spec}(x, ...)
}
\arguments{
\item{x}{A \code{\link{build_deep_spec}} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a deep-parameter taxonomy
}
