% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag-result.R
\name{print.dynhr_diagnostic}
\alias{print.dynhr_diagnostic}
\title{Print method for individual dynhr_diagnostic}
\usage{
\method{print}{dynhr_diagnostic}(x, ...)
}
\arguments{
\item{x}{A \code{dynhr_diagnostic} object.}

\item{...}{Currently unused.}
}
\description{
Print method for individual dynhr_diagnostic
}
