% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag-result.R
\name{print.dynhr_diagnostic_suite}
\alias{print.dynhr_diagnostic_suite}
\title{Print method for dynhr_diagnostic_suite}
\usage{
\method{print}{dynhr_diagnostic_suite}(x, ...)
}
\arguments{
\item{x}{A \code{dynhr_diagnostic_suite} object (list of diagnostics).}

\item{...}{Currently unused.}
}
\description{
Print method for dynhr_diagnostic_suite
}
