% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-passport.R
\name{print.dynhr_estimation_passport}
\alias{print.dynhr_estimation_passport}
\title{Print an estimation passport}
\usage{
\method{print}{dynhr_estimation_passport}(x, ...)
}
\arguments{
\item{x}{A \code{dynhr_estimation_passport}.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an estimation passport
}
