% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-full-estimation.R
\name{print.dynhr_estimation_result}
\alias{print.dynhr_estimation_result}
\title{Print an estimation result}
\usage{
\method{print}{dynhr_estimation_result}(x, ...)
}
\arguments{
\item{x}{A \code{dynhr_estimation_result} object}

\item{...}{Ignored}
}
\value{
\code{x} invisibly
}
\description{
Print an estimation result
}
