% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passport-run.R
\name{print.dynhr_passport_run}
\alias{print.dynhr_passport_run}
\title{Print an estimation-passport run report card}
\usage{
\method{print}{dynhr_passport_run}(x, ...)
}
\arguments{
\item{x}{A \code{dynhr_passport_run}.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an estimation-passport run report card
}
