% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramsey-augment-mod.R
\name{print.dynhr_ramsey_augmented}
\alias{print.dynhr_ramsey_augmented}
\title{Print method for dynhr_ramsey_augmented}
\usage{
\method{print}{dynhr_ramsey_augmented}(x, ...)
}
\arguments{
\item{x}{A \code{dynhr_ramsey_augmented} object.}

\item{...}{Unused; included for S3 compatibility.}
}
\description{
Print method for dynhr_ramsey_augmented
}
