% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mom-se-bounds.R
\name{print.dynhr_se_bounds}
\alias{print.dynhr_se_bounds}
\title{Print standard-error bounds}
\usage{
\method{print}{dynhr_se_bounds}(x, digits = 4L, ...)
}
\arguments{
\item{x}{A \code{"dynhr_se_bounds"} object from
\code{\link{mom_se_bounds}}.}

\item{digits}{Significant digits.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print standard-error bounds
}
