% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global-solution-methods.R
\name{print.GlobalSolution}
\alias{print.GlobalSolution}
\title{Print a GlobalSolution object}
\usage{
\method{print}{GlobalSolution}(x, ...)
}
\arguments{
\item{x}{A \code{GlobalSolution} object.}

\item{...}{Ignored.}
}
\description{
Print a GlobalSolution object
}
