% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hank-manifest2.R
\name{print.hank_het2_manifest}
\alias{print.hank_het2_manifest}
\title{Print a two-asset run manifest}
\usage{
\method{print}{hank_het2_manifest}(x, ...)
}
\arguments{
\item{x}{A \code{\link{hank_het2_manifest}}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a two-asset run manifest
}
\seealso{
\code{\link{hank_het2_manifest}}
}
