% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hank-manifest3.R
\name{print.hank_het3_manifest}
\alias{print.hank_het3_manifest}
\title{Print a three-asset run manifest}
\usage{
\method{print}{hank_het3_manifest}(x, ...)
}
\arguments{
\item{x}{A \code{\link{hank_het3_manifest}}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a three-asset run manifest
}
\seealso{
\code{\link{hank_het3_manifest}}
}
