% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hank-nonlinearity.R
\name{print.hank_nonlinearity_scan}
\alias{print.hank_nonlinearity_scan}
\title{Print a HANK nonlinearity scan}
\usage{
\method{print}{hank_nonlinearity_scan}(x, ...)
}
\arguments{
\item{x}{A \code{hank_nonlinearity_scan} from
\code{\link{hank_nonlinearity_scan}}.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a HANK nonlinearity scan
}
