% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp-estimand.R
\name{print.lp_estimand}
\alias{print.lp_estimand}
\title{Print an lp_estimand object}
\usage{
\method{print}{lp_estimand}(x, ...)
}
\arguments{
\item{x}{An \code{"lp_estimand"} object from \code{\link{lp_estimand}}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an lp_estimand object
}
