% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulant-likelihood.R
\name{sample_cumulants}
\alias{sample_cumulants}
\title{Compute sample (K-statistics) cumulants from observed data}
\usage{
sample_cumulants(Y, max_order = 4L)
}
\arguments{
\item{Y}{T × n data matrix}

\item{max_order}{Maximum cumulant order (2, 3, or 4; default 4)}
}
\value{
List:
  \item{mean}{n-vector of sample mean}
  \item{var_cov}{n × n sample variance-covariance}
  \item{c3}{n × n^2 sample third cumulant (K_3)}
  \item{c4}{n × n^3 sample fourth cumulant (K_4)}
  \item{n_obs}{number of observations}
}
\description{
Unbiased k-statistics for cumulants up to order 4, following the
classical definition (Fisher 1929).
}
