% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag-plot-theme.R
\name{scale_colour_dynhr_cividis}
\alias{scale_colour_dynhr_cividis}
\alias{scale_color_dynhr_cividis}
\title{Sequential cividis colour scale (viridis option "E")}
\usage{
scale_colour_dynhr_cividis(..., na.value = dynhr_na_colour)

scale_color_dynhr_cividis(..., na.value = dynhr_na_colour)
}
\arguments{
\item{...}{Additional arguments passed to
\code{ggplot2::scale_colour_viridis_c}.}

\item{na.value}{Colour for NA values (default: dynhr NA colour).}
}
\description{
Sequential cividis colour scale (viridis option "E")
}
