% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag-plot-theme.R
\name{scale_colour_dynhr_light}
\alias{scale_colour_dynhr_light}
\alias{scale_color_dynhr_light}
\title{Discrete line/colour scale using Paul Tol light in dynhr order}
\usage{
scale_colour_dynhr_light(..., na.value = dynhr_na_colour)

scale_color_dynhr_light(..., na.value = dynhr_na_colour)
}
\arguments{
\item{...}{Passed to \code{ggplot2::discrete_scale}.}

\item{na.value}{Colour for NA values (default: Tol grey).}
}
\description{
The colour counterpart of \code{\link{scale_fill_dynhr_light}} -- the pale
palette used for area fills, available on the \code{colour} aesthetic so an
outline can match its fill.  If there are more levels than the 6 light
colours, the palette is interpolated in Lab space.
}
