% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag-plot-theme.R
\name{scale_colour_dynhr_vibrant}
\alias{scale_colour_dynhr_vibrant}
\alias{scale_color_dynhr_vibrant}
\title{Discrete line/colour scale using Paul Tol vibrant in dynhr order}
\usage{
scale_colour_dynhr_vibrant(..., na.value = dynhr_na_colour)

scale_color_dynhr_vibrant(..., na.value = dynhr_na_colour)
}
\arguments{
\item{...}{Passed to \code{ggplot2::discrete_scale}.}

\item{na.value}{Colour for NA values (default: Tol grey).}
}
\description{
If there are more levels than the 6 vibrant colours, the palette is
interpolated in Lab space.
}
