% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag-plot-theme.R
\name{scale_fill_dynhr}
\alias{scale_fill_dynhr}
\title{Legacy: fill scale -- now Tol light in dynhr order}
\usage{
scale_fill_dynhr(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{scale_fill_dynhr_light}}.}
}
\description{
Legacy: fill scale -- now Tol light in dynhr order
}
