% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag-plot-theme.R
\name{scale_fill_dynhr_cividis}
\alias{scale_fill_dynhr_cividis}
\title{Sequential cividis fill scale (viridis option "E")}
\usage{
scale_fill_dynhr_cividis(..., na.value = dynhr_na_fill)
}
\arguments{
\item{...}{Additional arguments passed to
\code{ggplot2::scale_fill_viridis_c}.}

\item{na.value}{Fill colour for NA values (default: dynhr NA fill).}
}
\description{
Sequential cividis fill scale (viridis option "E")
}
