% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag-plot-theme.R
\name{scale_fill_dynhr_light}
\alias{scale_fill_dynhr_light}
\title{Discrete fill scale using Paul Tol light in dynhr order}
\usage{
scale_fill_dynhr_light(..., na.value = dynhr_na_fill)
}
\arguments{
\item{...}{Passed to \code{ggplot2::discrete_scale}.}

\item{na.value}{Fill for NA values (default: Tol pale grey).}
}
\description{
If there are more levels than the 6 light colours, the palette is
interpolated in Lab space.
}
