% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global-solution-methods.R
\name{simulate.GlobalSolution}
\alias{simulate.GlobalSolution}
\title{Stochastic simulation from the global policy function}
\usage{
\method{simulate}{GlobalSolution}(object, nsim = 100L, seed = NULL, init_state = NULL, ...)
}
\arguments{
\item{object}{A \code{GlobalSolution} object.}

\item{nsim}{Number of periods to simulate (default 100).}

\item{seed}{Random seed (default NULL = don't set).}

\item{init_state}{Named numeric; initial state lag values.
Defaults to steady state.}

\item{...}{Ignored.}
}
\value{
A matrix of size \code{nsim x n_endo} with simulated paths.
}
\description{
Stochastic simulation from the global policy function
}
