% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{solve_dsge}
\alias{solve_dsge}
\title{Solve a DSGE model (parse -> compile -> steady -> perturbation)}
\usage{
solve_dsge(...)
}
\arguments{
\item{...}{Arguments forwarded to \code{\link{solve_model}}.}
}
\value{
A \code{dynhr_solved} object.
}
\description{
Convenience alias for \code{\link{solve_model}}.
}
\seealso{
\code{\link{solve_model}}, \code{\link{run_mode_finding}}
}
