% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve-perturbation-order5.R
\name{solve_perturbation_order5}
\alias{solve_perturbation_order5}
\title{Solve the deterministic fifth-order perturbation of a DSGE model}
\usage{
solve_perturbation_order5(
  model,
  compiled,
  ss,
  params,
  dr4,
  h = 0.05,
  verbose = FALSE
)
}
\arguments{
\item{model}{dynhr_mod object from parse_mod()}

\item{compiled}{dynhr_compiled from compile_model()}

\item{ss}{Named numeric steady state vector}

\item{params}{Named numeric parameter vector}

\item{dr4}{Fourth-order DecisionRules4 object}

\item{h}{Step size for numerical derivatives (default 5e-2)}

\item{verbose}{Print progress}
}
\value{
A DecisionRules5 object extending DecisionRules4 with 5th-order
  fields ghxxxxx, ghxxxxu, ...
}
\description{
Given first- through fourth-order decision rules, computes the 5th-
order terms ghxxxxx, ghxxxxu, ... using Levintal (2017) compact
tensor notation.
}
\references{
Levintal, O. (2017). Fifth-order perturbation solution of DSGE models.
    \emph{Journal of Economic Dynamics and Control}, 80, 1-16.
  Andreasen, M. M., Fernandez-Villaverde, J., & Rubio-Ramirez, J. F. (2018).
    The pruned state-space system for non-linear DSGE models.
    \emph{Review of Economic Studies}, 85(1), 1-49.
}
