% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nash-ramsey.R
\name{summary.dynhr_nash_ramsey_openloop}
\alias{summary.dynhr_nash_ramsey_openloop}
\title{Summarise a Nash–Ramsey game result}
\usage{
\method{summary}{dynhr_nash_ramsey_openloop}(object, ...)
}
\arguments{
\item{object}{A \code{dynhr_nash_ramsey_openloop} object.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns the welfare matrix.
}
\description{
Summarise a Nash–Ramsey game result
}
