% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag-plot-theme.R
\name{tol_light}
\alias{tol_light}
\title{Paul Tol light palette, named (qualitative, suitable for area fills)}
\usage{
tol_light
}
\description{
Paul Tol light palette, named (qualitative, suitable for area fills)
}
