% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag-plot-theme.R
\name{tol_vibrant}
\alias{tol_vibrant}
\title{Paul Tol vibrant palette, named (qualitative, high-contrast)}
\usage{
tol_vibrant
}
\description{
Paul Tol vibrant palette, named (qualitative, high-contrast)
}
