% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{dynhr_get_options}
\alias{dynhr_get_options}
\title{Get all current dynhr package-level options}
\usage{
dynhr_get_options()
}
\value{
Named list of all currently set options.
}
\description{
Get all current dynhr package-level options
}
