% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hank-model.R
\name{hank_het_block_spec}
\alias{hank_het_block_spec}
\title{Wrap a heterogeneous-agent household as a sequence-space block}
\usage{
hank_het_block_spec(name, block, inputs = c("r", "w"), outputs = c("A", "C"))
}
\arguments{
\item{name}{Character block name.}

\item{block}{A \code{\link{hank_het_block}} solved at steady state.}

\item{inputs, outputs}{Character vectors; must be supported by
\code{\link{hank_het_jacobian}} (inputs a subset of \code{c("r","w")},
outputs a subset of \code{c("A","C")}).}
}
\value{
An object of class \code{hank_block} (kind \code{"het"}).
}
\description{
Wrap a heterogeneous-agent household as a sequence-space block
}
