% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hank-jacobian.R
\name{hank_het_jacobian}
\alias{hank_het_jacobian}
\title{Sequence-space Jacobian of a het block via the fake-news algorithm}
\usage{
hank_het_jacobian(
  block,
  T_h,
  inputs = c("r", "w"),
  outputs = c("A", "C"),
  delta_in = 1e-05,
  delta_va = 1e-06,
  delta_d = 1e-06
)
}
\arguments{
\item{block}{A \code{\link{hank_het_block}}.}

\item{T_h}{Integer horizon.}

\item{inputs}{Character subset of \code{c("r", "w")}.}

\item{outputs}{Character subset of \code{c("A", "C")}.}

\item{delta_in}{FD step for the input (r/w) perturbation in the s=0 term.}

\item{delta_va}{Relative FD step for the backward value-function propagation.}

\item{delta_d}{FD step for the distributional (curly-D) response.}
}
\value{
Nested list \code{J[[output]][[input]]}, each a \code{T x T} matrix
  with \code{[t, s] = dO_t/dI_s}.  (The date-0 response vectors are the first
  Jacobian row, \code{J[[o]][[i]][1, ]}.)
}
\description{
Sequence-space Jacobian of a het block via the fake-news algorithm
}
