% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochsimul-monolith.R
\name{print.StochSimulResult}
\alias{print.StochSimulResult}
\title{Print a stoch_simul result}
\usage{
\method{print}{StochSimulResult}(x, ...)
}
\arguments{
\item{x}{A `StochSimulResult` object from [stoch_simul()], [compute_irfs()] or [compute_moments()].}

\item{...}{Ignored.}
}
\value{
`x`, invisibly.
}
\description{
Print a stoch_simul result
}
