% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag-plot-theme.R
\name{scale_colour_dynhr}
\alias{scale_colour_dynhr}
\title{Legacy: colour scale -- now Tol vibrant in dynhr order}
\usage{
scale_colour_dynhr(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{scale_colour_dynhr_vibrant}}.}
}
\description{
Legacy: colour scale -- now Tol vibrant in dynhr order
}
