% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag-plot-theme.R
\name{scale_colour_dynhr_sunset}
\alias{scale_colour_dynhr_sunset}
\alias{scale_color_dynhr_sunset}
\title{Diverging colour scale using Paul Tol sunset (continuous)}
\usage{
scale_colour_dynhr_sunset(..., midpoint = 0, na.value = dynhr_na_colour)

scale_color_dynhr_sunset(..., midpoint = 0, na.value = dynhr_na_colour)
}
\arguments{
\item{...}{Passed to \code{ggplot2::scale_colour_gradientn}.}

\item{midpoint}{Numeric midpoint (default 0).}

\item{na.value}{Colour for NA values.}
}
\description{
Diverging colour scale using Paul Tol sunset (continuous)
}
